Separated Movable interface from Shooter
